Introduce an entry API which supersedes the existing API.
Transitions the implementation of most existing methods to be
based on entries, and exposes a flexible entry API, removing the
need for an infinite series of ever-increasing-in-complexity
combinators like `insert_with_opt` which can now be implemented
ergonomically outside of the crate.

Also renames the internal Entry type to Slot.